refactor: tools.py split Phase 4 — extract platform tools to _platforms.py
Moved Telegram/WeChat/Feishu/QQ setup tools (~448 lines) from _core.py
into _platforms.py.  Self-contained — only depends on paths + CONFIG_PATH.

_core.py now ~3600 lines (down from ~5269 originally — a 32% reduction).
All four phases complete: the original monolith is now split across 7
domain-specific submodules under sjtu_agent/agent/tools/:

  _reminders.py    _user_profile.py    _python_exec.py
  _email.py        _mcp_skills.py      _platforms.py
